Remove test_numpy_interop
